various

+++ scripts/record_demo.py
@@ -24,7 +24,20 @@ time_string  = time.strftime("%Y-%m-%d-%H-%M-%S", localtime)
 
 os.chdir(osp.dirname(args.master_file))
 
+if not osp.exists(args.master_file):
+    yn = yes_or_no("master file does not exist. create?")
+    basename = raw_input("what is the base name?\n").strip()
+    if yn:
+        with open(args.master_file,"w") as fh: fh.write("""
+name: %s
+h5path: %s
+bags:
+        """%(basename, basename+".h5"))
+    else:
+        print "exiting."
+        exit(0)
 with open(args.master_file, "r") as fh: master_info = yaml.load(fh)
+if master_info["bags"] is None: master_info["bags"] = []
 for suffix in itertools.chain("", (str(i) for i in itertools.count())):
     demo_name = args.demo_prefix + suffix
     if not any(bag["demo_name"] == demo_name for bag in master_info["bags"]):
@@ -68,8 +81,8 @@ if yes_or_no("save demo?"):
         fh.write("\n"
             "- bag_file: %(bagfilename)s\n"
             "  annotation_file: %(annfilename)s\n"
-            "  video_dir: %(videodir)s"
+            "  video_dir: %(videodir)s\n"
             "  demo_name: %(demoname)s"%dict(bagfilename=bagfilename, annfilename=annfilename, videodir=demo_name, demoname=demo_name))
 else:
-    shutil.rmtree(demo_name) #video dir
-    os.unlink(bagfilename)
+    if osp.exists(demo_name): shutil.rmtree(demo_name) #video dir
+    if osp.exists(bagfilename): os.unlink(bagfilename)
